Refactor: Extract canonical app entrypoint
